split up cli modules into submodules
