+++ nix_review/cli/__init__.py
@@ -1,81 +1,10 @@
 import argparse
 import os
-import re
-import subprocess
-import sys
-from contextlib import ExitStack
 from typing import Any, List
 
-from .buildenv import Buildenv
-from .review import CheckoutOption, Review, nix_shell
-from .worktree import Worktree
-
-
-def parse_pr_numbers(number_args: List[str]) -> List[int]:
-    prs: List[int] = []
-    for arg in number_args:
-        m = re.match(r"(\d+)-(\d+)", arg)
-        if m:
-            prs.extend(range(int(m.group(1)), int(m.group(2))))
-        else:
-            try:
-                prs.append(int(arg))
-            except ValueError:
-                print(f"expected number, got {m}", file=sys.stderr)
-                sys.exit(1)
-    return prs
-
-
-def pr_command(args: argparse.Namespace) -> None:
-    prs = parse_pr_numbers(args.number)
-    use_ofborg_eval = args.eval == "ofborg"
-    checkout_option = (
-        CheckoutOption.MERGE if args.checkout == "merge" else CheckoutOption.COMMIT
-    )
-
-    contexts = []
-
-    with ExitStack() as stack:
-        for pr in prs:
-            worktree = stack.enter_context(Worktree(f"pr-{pr}"))
-            try:
-                review = Review(
-                    worktree_dir=worktree.worktree_dir,
-                    build_args=args.build_args,
-                    api_token=args.token,
-                    use_ofborg_eval=use_ofborg_eval,
-                    only_packages=set(args.package),
-                    checkout=checkout_option,
-                )
-                contexts.append((pr, worktree, review.build_pr(pr)))
-            except subprocess.CalledProcessError:
-                print(
-                    f"https://github.com/NixOS/nixpkgs/pull/{pr} failed to build",
-                    file=sys.stderr,
-                )
-
-        for pr, worktree, attrs in contexts:
-            print(f"https://github.com/NixOS/nixpkgs/pull/{pr}")
-            os.environ["NIX_PATH"] = worktree.nixpkgs_path()
-            nix_shell(attrs)
-
-        if len(contexts) != len(prs):
-            sys.exit(1)
-
-
-def rev_command(args: argparse.Namespace) -> None:
-    commit = subprocess.run(
-        ["git", "rev-parse", "--verify", args.commit],
-        check=True,
-        stdout=subprocess.PIPE,
-    ).stdout.decode("utf-8")
-    with Worktree(f"rev-{commit}") as worktree:
-        review = Review(
-            worktree_dir=worktree.worktree_dir,
-            build_args=args.build_args,
-            only_packages=set(args.package),
-        )
-        review.review_commit(args.branch, commit)
+from ..buildenv import Buildenv
+from .pr import pr_command
+from .rev import rev_command
 
 
 def pr_flags(subparsers: argparse._SubParsersAction) -> argparse.ArgumentParser:
